# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Bal Suraksha API
  version: "0.1.0"
  description: |
    Private child-safety support platform. All endpoints enforce organization-scoped
    authorization. Enums are fixed; unknown values are rejected server-side.
  contact:
    name: Bal Suraksha engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development

# ── Security schemes ────────────────────────────────────────────────────────
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Short-lived JWT issued on private session creation.
    staffToken:
      type: http
      scheme: bearer
      description: JWT issued on staff login (local stub for demo; OIDC in production).

  # ── Shared error envelope ─────────────────────────────────────────────────
  schemas:
    ErrorEnvelope:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message. Never contains private case content.
        code:
          type: string
          description: Machine-readable error code.
        request_id:
          type: string

    # ── Enums ─────────────────────────────────────────────────────────────
    CaseStatus:
      type: string
      enum:
        - RECEIVED
        - TRIAGE
        - ASSIGNED
        - ACCEPTED
        - IN_PROGRESS
        - FOLLOW_UP
        - CLOSED
    AlertStatus:
      type: string
      enum:
        - DRAFT
        - IN_REVIEW
        - APPROVED
        - ACTIVE
        - RESOLVED
        - WITHDRAWN
        - EXPIRED
    NotificationMode:
      type: string
      enum:
        - DISABLED
        - TEST_ALLOWLIST
        - LIVE
...
    SessionCreateRequest:
      type: object
      required: [mode]
      properties:
        entry_point_id:
          type: string
          format: uuid
        mode:
          type: string
          enum: [one_time, with_return_access]
        language:
          type: string
          default: en

        invitation_code:
          type: string
          description: Required for invited beta intake.
    SessionCreateResponse:
      type: object
      required: [session_token, expires_at]
      properties:
        session_token:
          type: string
          description: Short-lived JWT. Not the return secret.
        expires_at:
          type: string
          format: date-time
        session_id:
          type: string
          format: uuid
        return_code:
          type: string
          description: Random 256-bit hexadecimal capability, shown only at session creation. Save privately.
        return_expires_at:
          type: string
          format: date-time
    SessionAccessRequest:
      type: object
      required: [return_code]
      properties:
        return_code:
          type: string
          description: 64 hexadecimal characters. Display separators and whitespace are accepted. Legacy words are not accepted.
    SessionAccessResponse:
      type: object
      required: [session_token, case_id]
      properties:
        session_token:
          type: string
        case_id:
          type: string
          format: uuid

    # ── Assessment (AI) ───────────────────────────────────────────────────
    AssessmentRequest:
      type: object
      required: [selected_text, language]
      properties:
        selected_text:
          type: string
          maxLength: 5000
          description: Text selected by the user. Treated as untrusted data.
        language:
          type: string
          description: IETF language tag, e.g. en, hi.
    AssessmentResponse:
      type: object
      required: [status, human_help_available]
      properties:
        status:
          type: string
          enum: [ok, unavailable, unsupported_language, input_too_long]
        human_help_available:
          type: boolean
          description: Always true. Human help path is never blocked by AI status.
        observed_behaviors:
          type: array
          items:
            type: string
          description: Specific behaviors visible in the submitted text. Grounded only.
          nullable: true
        uncertainty_statement:
          type: string
          description: Explicit acknowledgement of context limits.
          nullable: true
        clarifying_questions:
          type: array
          items:
            type: string
          maxItems: 3
          nullable: true
        suggested_options:
          type: array
          items:
            type: string
          nullable: true
        provider:
          type: string
          description: "stub | gemini-2.5-flash (or actual model used)"
        is_simulated:
          type: boolean

    # ── Cases ─────────────────────────────────────────────────────────────
    CaseCreateRequest:
      type: object
      required: [account_text]
      properties:
        account_text:
          type: string
          maxLength: 10000
        safe_contact_preference:
          $ref: "#/components/schemas/SafeContactPreference"
        conflict_flag:
          type: string
          nullable: true
          enum: [school_implicated, caregiver_implicated, staff_implicated]
        optional_safe_contact:
          type: string
          nullable: true
        route_type:
          type: string
          enum: [ask_for_help, worried_about_someone, missing_child]
    SafeContactPreference:
      type: object
      properties:
        time_zone:
          type: string
          enum: [Asia/Kolkata, UTC]
          default: Asia/Kolkata
        safe_hours_start:
          type: string
          nullable: true
        safe_hours_end:
          type: string
          nullable: true
        preferred_channel:
          type: string
          enum: [message_in_app, no_contact]
          nullable: true
    CaseCreateResponse:
      type: object
      required: [case_id, receipt_id, status]
      properties:
        case_id:
          type: string
          format: uuid
        receipt_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        received_at:
          type: string
          format: date-time
...
    CaseStaffView:
      type: object
      description: Responder/staff view. Restricted to assigned organization.
      required: [case_id, status, created_at, route_type]
      properties:
        case_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        route_type:
          type: string
        conflict_flag:
          type: string
          nullable: true
        account_text:
          type: string
        safe_contact_preference:
          $ref: "#/components/schemas/SafeContactPreference"
          nullable: true
        created_at:
          type: string
          format: date-time
        assigned_to:
          type: string
          format: uuid
          nullable: true
        assignment_accepted_at:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer

    # ── Assignments ───────────────────────────────────────────────────────
...
    MessageSendRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 5000

    # ── Alerts ────────────────────────────────────────────────────────────
    AlertDraftRequest:
      type: object
      required: [case_id, description_text, issuer_name, area_id, expiry_at, verification_reference]
      properties:
        case_id:
          type: string
          format: uuid
        description_text:
          type: string
          maxLength: 500
          description: Approved public description only. No private narrative.
        issuer_name:
          type: string
        area_id:
          type: string
          description: Active registered area ID from subscriptions/config; exact area matching, not GPS tracking.
        expiry_at:
          type: string
          format: date-time
        verification_reference:
          type: string
          maxLength: 500
          description: Staff-only reference supporting verification. Tips remain in-app.
    AlertRevision:
      type: object
      required: [alert_id, status]
      properties:
        revision_id:
          type: string
          format: uuid
        alert_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AlertStatus"
        description_text:
          type: string
        issuer_name:
          type: string
        coverage_polygon:
          type: object
        expiry_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        approved_by:
          type: string
          format: uuid
          nullable: true
        approved_at:
          type: string
          format: date-time
          nullable: true
    AlertApprovalRequest:
      type: object
      required: [revision_id, decision]
      properties:
        revision_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approve, reject]
        notes:
          type: string
          nullable: true
    PublicAlertView:
      type: object
      description: |
        Approved public projection only. No private case data, no sighting details,
        no family phone numbers, no precise coordinates.
      required: [alert_id, revision_id, status, description_text, issuer_name, expiry_at, current_status_url]
      properties:
        alert_id:
          type: string
          format: uuid
        revision_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AlertStatus"
        description_text:
          type: string
        issuer_name:
          type: string
        last_seen_area:
          type: string
          nullable: true
        expiry_at:
          type: string
          format: date-time
        issued_at:
          type: string
          format: date-time
        tip_route_available:
          type: boolean
        current_status_url:
          type: string
          format: uri
        is_test:
          type: boolean
          description: True in demo/TEST_ALLOWLIST mode. Label must be shown in UI.

    # ── Subscriptions ─────────────────────────────────────────────────────
    SubscriptionRequest:
      type: object
      required: [endpoint, auth, p256dh, area_id, consent_confirmed, management_secret]
      properties:
        endpoint:
          type: string
          format: uri
          description: HTTPS PushSubscription endpoint on an allowed browser push service.
        auth:
          type: string
          description: Base64url 16-byte browser subscription authentication secret.
        p256dh:
          type: string
          description: Base64url valid P-256 browser subscription public key.
        area_id:
          type: string
        management_secret:
          type: string
          description: Client-generated 32 random bytes in unpadded base64url; retained by this browser and only a SHA-256 digest stored by the API.
        enrollment_code:
          type: string
          description: Required organizer invitation in TEST mode. Mode is selected by the server, never the requester.
        language:
          type: string
          default: en
        consent_confirmed:
          type: boolean
          description: Must be true. User must have seen the consent explanation.

    # ── Tips ─────────────────────────────────────────────────────────────
    TipCreateRequest:
      type: object
      required: [alert_id, sighting_description]
      properties:
        alert_id:
          type: string
          format: uuid
        sighting_description:
          type: string
          maxLength: 5000
        approximate_location:
          type: string
          nullable: true
        contact_preference:
          type: string
          nullable: true
    TipCreateResponse:
      type: object
      required: [receipt_id, status]
      properties:
        receipt_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [received]

    # ── Health ────────────────────────────────────────────────────────────
...
paths:
  /subscriptions/config:
    get:
      operationId: notificationConfiguration
      summary: Registered areas, public VAPID key and enrollment readiness
      responses:
        "200":
          description: Public notification configuration; no private VAPID key or invitation is returned
          content:
            application/json:
              schema:
                type: object
                required: [areas, is_test, mode, enrollment_available, public_key]
                properties:
                  areas:
                    type: array
                    items:
                      type: object
                      required: [id, name]
                      properties:
                        id: {type: string}
                        name: {type: string}
                  is_test: {type: boolean}
                  mode: {$ref: '#/components/schemas/NotificationMode'}
                  enrollment_available: {type: boolean}
                  public_key: {type: string}
        "503":
          description: Area registration is unavailable
  /public/alerts:
    get:
      operationId: listPublicAlerts
      summary: Up to 50 current alerts for the runtime environment, optionally filtered by registered area
      parameters:
        - name: area_id
          in: query
          schema: {type: string}
      responses:
        "200":
          description: Current approved public projections; no private case fields
          content:
            application/json:
              schema:
                type: object
                required: [alerts, is_test]
                properties:
                  is_test: {type: boolean}
                  alerts:
                    type: array
                    items: {$ref: '#/components/schemas/PublicAlertView'}
...
  /sessions:
    post:
      operationId: createSession
      summary: Start a private session
      tags: [sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreateRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreateResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /sessions/{session_id}:
    delete:
      operationId: endSession
      summary: Explicitly end a session (Incognito exit)
      tags: [sessions]
      security:
        - sessionToken: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session ended. Credential revoked. Submitted case is not deleted.
        "401":
          description: Invalid or expired session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /session-access:
    post:
      operationId: accessByReturnSecret
      summary: Retrieve case access using return secret words
      tags: [sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionAccessRequest"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionAccessResponse"
        "401":
          description: Invalid secret or rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Assessments (AI)
  /assessments:
    post:
      operationId: createAssessment
      summary: Request bounded AI assessment of selected text
      tags: [assessments]
      security:
        - sessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentRequest"
      responses:
        "200":
          description: Assessment result (or unavailable)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"

  # Cases
  /cases:
    post:
      operationId: createCase
      summary: Submit a support request
      tags: [cases]
      security:
        - sessionToken: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: Client-generated UUID. Retry with same key returns original receipt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCreateRequest"
      responses:
        "201":
          description: Case created and durably received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCreateResponse"
        "200":
          description: Existing case returned (idempotent retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCreateResponse"
        "409":
          description: Same idempotency key with different payload rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /staff/cases:
    get:
      operationId: listStaffCases
      summary: List cases for the authenticated staff member's organization
      tags: [staff-cases]
      security:
        - staffToken: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of cases (org-scoped)
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/CaseStaffView"
                  total:
                    type: integer
...
  /staff/cases/{case_id}/messages:
    post:
      operationId: sendMessage
      summary: Send a message on a case (staff or authorized session)
      tags: [messages]
      security:
        - staffToken: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageSendRequest"
      responses:
        "201":
          description: Message sent

  # Staff auth: configured Keycloak OIDC, or local passwords in demo mode only.
...
  /staff/auth/me:
    get:
      operationId: staffAuthMe
      summary: Current active application staff mapping
      tags: [auth]
      security:
        - staffToken: []
      responses:
        "200":
          description: Active application identity and role; response is not cached
          content:
            application/json:
              schema:
                type: object
                required: [staff_id, organization_id, role]
                properties:
                  staff_id:
                    type: string
                    format: uuid
                  organization_id:
                    type: string
                    format: uuid
                  role:
                    type: string
                    enum: [responder, supervisor, alert_preparer, alert_approver, admin]
        "401":
          description: Invalid or expired staff access token
        "403":
          description: No active application staff access
  /staff/auth/login:
    post:
      operationId: staffLogin
      summary: Staff login (local stub for demo)
      description: Rejected outside demo mode or when STAFF_AUTH_MODE is oidc. Organization sign-in uses the identity provider directly.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                required: [token, expires_at, staff_id, role]
                properties:
                  token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                  staff_id:
                    type: string
                    format: uuid
                  role:
                    type: string
                    enum: [responder, supervisor, alert_preparer, alert_approver]
                  organization_id:
                    type: string
                    format: uuid
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Alerts (staff)
  /staff/alert-drafts:
    post:
      operationId: createAlertDraft
      summary: Create an alert draft (alert_preparer role required)
      tags: [alerts]
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertDraftRequest"
      responses:
        "201":
          description: Alert draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRevision"
...
  /staff/alert-approvals:
    post:
      operationId: submitAlertApproval
      summary: Approve or reject an alert revision (alert_approver role; cannot self-approve)
      tags: [alerts]
      security:
        - staffToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertApprovalRequest"
      responses:
        "200":
          description: Decision recorded
        "403":
          description: Self-approval rejected or insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Revision has changed since last seen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /subscriptions:
    post:
      operationId: createSubscription
      summary: Subscribe to alerts for an area (explicit consent required)
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                type: object
                required: [subscription_id]
                properties:
                  subscription_id:
                    type: string
                    format: uuid
        "422":
          description: Allowlist full or consent not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /tips:
    post:
      operationId: createTip
      summary: Submit a private sighting tip (public — opaque receipt returned)
      tags: [tips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TipCreateRequest"
      responses:
        "201":
          description: Tip received (opaque receipt only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TipCreateResponse"
  /staff/tips:
    get:
      operationId: listTips
      summary: List tips for an active alert (authorized reviewers only)
      tags: [tips]
      security:
        - staffToken: []
      parameters:
        - name: alert_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tips list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tips:
                    type: array
                    items:
                      type: object
                      required: [tip_id, sighting_description, received_at]
                      properties:
                        tip_id:
                          type: string
                          format: uuid
                        sighting_description:
                          type: string
                        approximate_location:
                          type: string
                          nullable: true
                        received_at:
                          type: string
                          format: date-time